images - refactoring part 1
